renamed textUtils.jsx to textAnalysisUtils.jsx
