+++ src/components/NavItems/Assistant/utils/textAnalysisUtils.jsx
